+++ src/main/java/ch/bfh/unicrypt/math/algebra/additive/abstracts/AbstractAdditiveMonoid.java
@@ -97,7 +97,7 @@ public abstract class AbstractAdditiveMonoid<E extends AdditiveElement<V>, V ext
 	}
 
 	@Override
-	public final E times(final Element element, final int amount) {
+	public final E times(final Element element, final long amount) {
 		return this.selfApply(element, amount);
 	}
