+++ src/main/java/ch/bfh/unicrypt/math/algebra/multiplicative/abstracts/AbstractMultiplicativeCyclicGroup.java
@@ -97,7 +97,7 @@ public abstract class AbstractMultiplicativeCyclicGroup<E extends Multiplicative
 	}
 
 	@Override
-	public final E power(final Element element, final int amount) {
+	public final E power(final Element element, final long amount) {
 		return this.selfApply(element, amount);
 	}
