+++ src/main/java/ch/bfh/unicrypt/math/algebra/multiplicative/abstracts/AbstractMultiplicativeGroup.java
@@ -95,7 +95,7 @@ public abstract class AbstractMultiplicativeGroup<E extends MultiplicativeElemen
 	}
 
 	@Override
-	public final E power(final Element element, final int amount) {
+	public final E power(final Element element, final long amount) {
 		return this.selfApply(element, amount);
 	}
